server02 - first producer consumer problem almost done
serverConnect02 can accept  multiple clients over the open socket. Threads wake up as clients connect and sleep when not needed but the conditions are not perfect. I'm not sure thread sleep or quit when the escape key it pressed and I need to find a way for the main thread to know when to quit. Once I fine tune this client to thread problem, I can implement the same logic for the logger thread.