It works
minor quirks need to be fixed such as just pressing enter does count as spelt wrong but maybe the thread should notice that.

I have tested the spell checker server with multiple clients on the single socket,  the threads service a client until 'esc' is pressed and then goes onto the next client or waits. the server quits when a thread reads '?'. The logger takes jobs as they appear and waits until signaled. 

I was getting a seg fault with my createLogEntry function. I was trying to store a string in a space that was big enough. malloc'ing space for the string seemed to solve that. a print statement was also causing another seg fault but it was a test statement so I could be deleted. I isolated my code line by line by staring lines in and out and "here#" print statements to find where exactly the fault was.