Finsihed
spellChechServer opens a port and allows 3 users + 3 waiters at any one time. when i client presses esc, the thread drops the client and picks up another. When ? is entered, the program exits.
server will connect to with default port number 1234 and use the dictionary file words.
if you want to use your own dictionary file, you need to specify a port number as well. 

I fixed a minor issue with the logger reporting the wrong correctness and with how the server acts if no port number is given